requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
pillow==10.3.0
streamlit==1.35.0
google-api-python-client==2.130.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
pandas==2.2.2
